:art: Refactored code to improve maintainability (#97)
* :arrow_up: Updated dependencies

Updated aws-java-sdk to 1.11.277

* :art: Created helper classes

Created AwsFileHelper for file operations
Created AwsConfigHelper for config file operations
Created AwsCredentialsHelper for credentials file operations
Created AwsSessionHelper for session management
Created Session model

Resolves: IOException due to the .aws folder not existing, nor being created by the code.

* :bug: Fixed tests bug
Fixed a bug with tests running on Windows machines, due to new line differences (\r\n as opposed to \n)

Attached new artifact

* :art: Moved authentication to a separate class

Created OktaAuthentication for Primary Authentication handling
Created OktaMFA for Secondary (2FA) Authentication handling
Create MenuHelper for prompting the user for a menu reply

* :art: Moved Okta SAML handling

Created OktaSaml to handle Okta SAML responses for the AWS app

* :bug: Fixed minor bug with file paths
Fixed AwsConfigHelper to point to the correct path
Fixed AwsCredentialsHelper to point to the correct path
Fixed AwsSessionHelper to point to the correct path

Reverted config.properties from previous commit

Attached new artifact

* :art: Completed refactoring for authentication and MFA

Added handleTimeoutsAndChanges to remove duplicate code in OktaMFA
Added proper descriptions for all methods in OktaAuthentication and OktaMFA

* :art: Refactored MultipleProfile to match project

Completely merged MultipleProfile
Refactored code to match project

* :art: Refactored the remaining code
:fire: Removed unused imports

* :bug: Fixed an issue with files

* Merged 06989d2:

* :art: Moved Profile.class to Models

* Created OktaAwsCliEnvironment to contain environment variables

* :art: Reformated code and optimised imports

* :art: Removed unnecessary attributions

* :bug: Fixed path building
:bug: Moved the .okta folder back to USER_HOME
:bug: Fixed bug with logout

* Reverted the default config.properties file

* :bug: Fixed the credential reprompt system

* :art: Made OktaAwsCliEnvironment be a constructor parameter for most classes, making them not static
:bug: Fixed bug in logging out multiple profiles due to FileReader not being closed

* :art: Optimized imports and formatted code

* :bug: Fixed multiple profiles not being used
:art: Added oktaawscli.jar to the .gitignore file
