+++ .gitignore
@@ -36,3 +36,6 @@ target/classes/log4j2.xml
 dependency-reduced-pom.xml
 .idea
 *.iml
+move.bat
+
+out/oktaawscli.jar
