+++ config.properties
@@ -18,3 +18,4 @@ OKTA_AWS_APP_URL=https://acmecorp.oktapreview.com/home/amazon_aws/0oa5zrwfs815KJ
 OKTA_USERNAME=
 OKTA_PASSWORD=
 OKTA_AWS_ROLE_TO_ASSUME=
+OKTA_PROFILE=
